Now using FragmentPagerAdapter and ViewPager.
As I said, I will change everything I spent hours doing. Hate my
life.

Now the FragmentChangeController extends FragmentPagerAdapter which is a
better way to store and switch Fragments apparently. At least I think it
is because it was made by smart people I assume.
Also, the fragments are now displayed in the ViewPager. This gives me
the ability to link it to the adapter and it adds a cool slide effect.
Maybe I will also be able to switch the ViewPagerto something cooler
later since I have an adapter set up already.
Overall, 5/10 experience because of wasted but pretty happy with the way
it turned out.