Finished up the barcode scanner. Camera is now full screen and recognizes the barcodes and draws an overlay around them.
Fix:
Make camera start with fragment not on click. Problem with layout width
and height cause it is talking it from layout that is not finished
building so the size ends up being 0, 0.
Camera only sees 3 tags at a time. Or at least draws around three tags.
Fix that too